requirements.txt
torch
torchvision
diffusers
accelerate
transformers
Pillow
numpy
scikit-learn
matplotlib